sklep dodaje produkty do bazy; uzytkownik moze je kupic jezeli asortyment jest dostepny; stworzenie obslugi uzytkownika w konsoli

+++ src/main/java/shop/orm/menagers/PurchaseManager.java
@@ -15,7 +15,7 @@ public class PurchaseManager {
         this.purchaseRepository = new PurchaseRepository();
     }
 
-    public List<Purchase> getAllPurchasesByClient(EntityManager entityManager, int clientId) {
+    public List<Purchase> getAllPurchasesByClient(EntityManager entityManager, long clientId) {
         return purchaseRepository.getAllPurchasesByClient(entityManager, clientId);
     }
